requirements.txt
google-cloud-bigquery==2.33.0
Flask==2.0.2